Macro shows demo roar icon when disabled
This way there's a visual difference between these two states: "Unable to buff due to combat" and "unable to buff because there is nothing to do".
The question mark icon is only visible when there is something to do but combat is preventing it. Otherwise the icon will be the specific buff or the DISABLED icon (demo roar).